[AAP-41483] feat: add edited_by to activation and allow is_enabled to be edited (#1232)
<!-- Mandatory: Provide a clear, concise description of the changes and
their purpose -->
The change allows user to directly start the activation after it is
updated by one API call.
It also shows who updates the activation by the edited_by field. It is
to be used together with the already existing edited_at field.
<!-- If applicable, provide a link to the issue that is being addressed
-->
[AAP-41483](https://issues.redhat.com/browse/AAP-41483)

<!-- What is being changed? -->
Added edited_by field to model Activation.
Added is_enabled field to the editable list in the serializer
<!-- Why is this change needed? -->
<!-- How does this change address the issue? -->
<!-- Does this change introduce any new dependencies, blockers or
breaking changes? -->
<!-- How it can be tested? -->
To test the new feature, set is_enabled to true in the patch API body
and verify the activation is restarted after it is updated. The patch
API returns the updated activation includes the new field edited_by.