fix: update credentials in an activation (#1251)
<!-- Mandatory: Provide a clear, concise description of the changes and
their purpose -->
Rework how to update credentials and extra_vars in an activation
<!-- If applicable, provide a link to the issue that is being addressed
-->
[AAP-42639](https://issues.redhat.com/browse/AAP-42639)
<!-- What is being changed? -->
If an activation contains credentials that have extra_vars injector in
their type schemas, for example, the credential automatically attached
when an event_stream is used, the code will automatically insert extra
vars to the activation's extra_var field. When such credentials are
added/removed/updated in the activation, the corresponding extra_var is
not updated properly.

To fix the problem we now restore the extra_var to its origin before
credential insertion. Then re-insert additional extra vars to this
attribute one by one iterating through new credentials in the
activation. In order to work properly for all possible updates, we pull
all required attributes from the DB if they are not given by the update
API.

<!-- How it can be tested? -->
To test the fix try to add/remove/update the credentials with extra_vars
injector or event_stream and verify the extra_var field in the
activation is updated correctly