feat(dispatcherd): add advisory lock for task uniqueness (#1262)
For the PoC of dispatcherd we used [pg
advisory_locks](https://www.postgresql.org/docs/current/explicit-locking.html#ADVISORY-LOCKS)
to ensure job uniqueness. It will be part of the final implementation.
It has been so far a robust and simpler way to ensure uniqueness in
comparison to existing `unique_enqueue` helper, which would not have a
replacement.

Since the use of the advisory lock is not tied to dispatcherd, this PR
explores its use as well for the current implementation with RQ redis,
assuming that will simplify the later implementation under a feature
flag of the dispatcherd itself.

There is no functional change and everything should work transparently.
I have been testing it extensively and tomorrow will do more battery of
tests, I have confidence. I leave the open the draft so the rest of the
team can look at the code as well as perform their own tests if you
want.

Jira: https://issues.redhat.com/browse/AAP-43538

---------

Signed-off-by: Alex <aizquier@redhat.com>
Co-authored-by: Bill Wei <bilwei@redhat.com>