ci: add param for dispatcherd flag in e2e-dispatch (#1294)
Add a new param to set the dispatcherd feature flag in our e2e-dispatch
workflow.
This workflow is always executed from main. This allows us to run on
demand the test suite with the feature flag enabled. Once the feature
flag is implemented and merged in main, we will have to update this
workflow to add a matrix strategy and run the test suite by default with
the flag enabled. For now, it is intended to work only for on-demand
runs.

Signed-off-by: Alex <aizquier@redhat.com>