refactor: get feature flags from features module (#1293)
<!-- Mandatory: Provide a clear, concise description of the changes and
their purpose -->
Get whether a feature flag is on or off as a constant from the new
features module.
It improves the code readability and slightly the performance as the
flags are now cached.

This work is to prepare the merging of the dispatcherd feature which is
guarded by a flag.