Bug/remove container-aap 40077 (#1284)
Objectives for this bug [https://issues.redhat.com/browse/AAP-40077]:
- Update restart API endpoint such that, when a user attempts to
restart, we detect if the deployment type is 'podman' and the
'activation-worker-node' is offline, we should block the restart and
return an error warning the user that a running container will be
orphaned.
- Enhance restart API endpoint with a force option where, if enabled,
the restart will go through.

<!-- Why is this change needed? -->
Activation containers are being left behind (orphaned) on nodes that go
offline in an multi-node deployment.

<!-- Why is this change needed? -->
The user has no way of knowing they they are leaving behind these
orphaned containers.

<!-- How it can be tested? -->