ci: upgrade ruff (#1325)
Ruff was added at the beginning of the project in an experimental way
but it has demonstrated to be a robust, performant and great tool. Now
it is quitely outdated.
This is the minimal change required to upgrade it. I want to iterate
over its configuration to provide more coverage.
Includes:
- Upgrade ruff dependency in poetry
- Update the syntax of commands and configuration
- Refactor some parts of the code that would raise lint errors with the
current config - type-comparison (E721)

Signed-off-by: Alex <aizquier@redhat.com>