fix: ensure rulebook_queue_name is not sanitized outside dispatcherd flag (#1351)
We have found a bug where the installers set a rulebook queue_name
longer than 63 chars, a limit imposed for pg_notify (dispatcherd feature
flag). This was set at the initialization of settings, where we can not
check yet the status of the flag.
This PR ensure the queue name is not sanitized for pg_notify when the
feature flag is not enabled.

[Jira: AAP-48235](https://issues.redhat.com/browse/AAP-48235)

Signed-off-by: Alex <aizquier@redhat.com>