AAP-48392 Fix test failures due to uptream DAB ContentType changes (#1355)
This is another version of
https://github.com/ansible/eda-server/pull/1353, but with the dependency
changes reverted.

That linked PR (draft) shows that this is effective. I expect tests will
pass here (which current DAB) as well. But that PR shows that, if we
merge this, it will make downstream tests for
https://github.com/ansible/django-ansible-base/pull/749 will pass.

----

Technical summary:

There are 2 structural changes in DAB RBAC that require adjustment by
the app:
- the `post_migrate` signal will now run _once_ as opposed to run _for
every app_ and to do that, it can only run when the post migrate signal
is called for its own app (the "dab_rbac" app), but doing this mucks
with the assumptions around what order post_migrate methods run in, so
this often requires other post-migrate methods to call the methods to
create DAB types and permissions to resolve the ordering problem
- A DAB RBAC-specific content type app is introduced, and this is
clearly not the same as the proper `ContentType` model, and this will
error any queries that pass a content type object as a python object. To
do that, we'll just use the variable from DAB RBAC for the content type,
which will give the correct model for whatever version of DAB we are
using.