AAP-48412 Unified RBAC display, sync with resource server (#1363)
This is the supporting eda-server patch for the core backend change for
ASTRAT-1464

requires https://github.com/ansible/django-ansible-base/pull/749

This adds endpoints and enables syncing logic to mirror EDA permissions
in the resource server, so that a unified view of all the permissions to
objects can be given. A recap of the general component changes, which
are being applied to eda-server here:

 - register role definition in the resource registry so it can be synced
- disable sync of role definitions when creating initial data because
sync isn't possible when that runs
 - enable new /service-index/ endpoints from DAB
 - add migration dependencies

The migration dependencies were done slightly different here, because
the testing of the "Project Admin" role rename required migrating
forward & backwards, and I had no intent of making the DAB migrations
reversible. It seems to work fine as it is here.

Testing is mostly with aap-dev.