fix: [AAP-52948] Enforce user input type checking in credentials module (#1403)
With this change, the backend will throw an error when a user inputs
values that are not booleans or strings when using the credentials
module.

The Jira ticket for this issue is:
https://issues.redhat.com/browse/AAP-52948

Three files were changed:

- `src/core/utils/credentials.py`
- `tests/integration/api/test_activation_with_credential.py`
- `tests/unit/test_credential_validation.py`

The changes to credentials.py implement the type-checking for user
inputs. I made minor changes to `test_activation_with_credential.py`
because some test values were provided as integers but according to the
new check, they should be strings. This change should not impact those
tests.

The changes to `test_credenial_validation.py` were added to test that
the server sends a proper error message when an integer is provided
instead of a string.

This change addresses the issue by enforcing type-checking early and
providing a more helpful error message about the problem.

There are no braking changes with this PR.